requirements.txt
requests==2.19.1
PyGithub==1.43
click==6.7
colorama==0.3.9
